Move function, remove repaint_needed
